Add files via upload
This application is made using  Kivy .Kivy is a Open source Python library for rapid development of applications
that make use of innovative user interfaces, such as multi-touch apps. Kivy runs on Linux, Windows, OS X, Android, iOS, and Raspberry Pi.
This is a application which is used for choosing appropriate gift for a particular occasion.
First the user has to select which kind of occasion the user has to attend and then the user will be given various options to choose from.
After the user has selected a option then he has to click on the particular button .Then the button will redirect to a website on which the user can buy the gift.